# docs/swagger.yaml
definitions:
...
  controllers.ConfrimationInput:
    properties:
      description:
        type: string
      imageUrl:
        type: string
      orderId:
        type: integer
      transferAmount:
        type: integer
    type: object
...
  controllers.OrderInput:
    properties:
      address:
        type: string
      city:
        type: string
      id:
        type: integer
      phoneNumber:
        type: string
      recipientsName:
        type: string
      statusPayment:
        type: string
    type: object
...
  models.Confrimation:
    properties:
      createdAt:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      orderId:
        type: integer
      transferAmount:
        type: integer
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  models.Order:
    properties:
      address:
        type: string
      city:
        type: string
      confrimations:
        items:
          $ref: '#/definitions/models.Confrimation'
        type: array
      createdAt:
        type: string
      id:
        type: integer
      orderDate:
        type: string
      orderDetails:
        items:
          $ref: '#/definitions/models.OrderDetail'
        type: array
      phoneNumber:
        type: string
      recipientsName:
        type: string
      statusPayment:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  models.OrderDetail:
    properties:
      createdAT:
        type: string
      harga:
        type: integer
      id:
        type: integer
      jumlah:
        type: integer
      orderId:
        type: integer
      productId:
        type: integer
      updatedAT:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /confrimation:
    get:
      description: Get list Confrimations refrence By userID
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Confrimation'
      security:
      - BearerToken: []
      summary: Get Confrimations By User Id
      tags:
      - Confrimation
    post:
      description: Creating a new Confrimation.
      parameters:
      - description: the body to create a new Confrimation
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ConfrimationInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Confrimation'
      security:
      - BearerToken: []
      summary: Create New Confrimation.
      tags:
      - Confrimation
  /confrimation/{id}:
    delete:
      description: Delete a Confrimation by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Confrimation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Confrimation.
      tags:
      - Confrimation
    patch:
      description: Update Confrimation by id.
      parameters:
      - description: the body to update a new Confrimation
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ConfrimationInput'
      - description: Confrimation id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Confrimation'
      security:
      - BearerToken: []
      summary: Update Confrimation.
      tags:
      - Confrimation
  /confrimation/approve/{id}:
    get:
      description: Approve Confrimation Payment order only Role admin.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Confrimation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Approve Confrimation Payment.
      tags:
      - Confrimation
...
  /orders:
    get:
      description: Get list orders refrence By userID
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      security:
      - BearerToken: []
      summary: Get Order By User Id
      tags:
      - Order
    post:
      description: Creating a new Order.
      parameters:
      - description: the body to create a new Order
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.OrderInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      security:
      - BearerToken: []
      summary: Create New Order.
      tags:
      - Order
  /orders/{id}:
    delete:
      description: Delete a Order by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Order.
      tags:
      - Order
    patch:
      description: Update Order by id.
      parameters:
      - description: the body to update a new Order
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.OrderInput'
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      security:
      - BearerToken: []
      summary: Update Order.
      tags:
      - Order
...
swagger: "2.0"